notification/utils.py
from asgiref.sync import async_to_sync
from channels.layers import get_channel_layer

def send_notification(user, message):
    channel_layer = get_channel_layer()

    async_to_sync(channel_layer.group_send)(
        f"user_{user.id}",
        {
            "type": "send_notification",
            "message": message
        }
    )